added ObjectRemover class; moved a couple things around
