+++ decepticon/tests/test_util.py
@@ -1,9 +1,30 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
+import tensorflow as tf
 from PIL import Image
 
-from decepticon._util import _load_to_array
+tf.enable_eager_execution()
+
+from decepticon._util import _load_to_array, _remove_objects
+
+
+"""
+
+        BUILDING TEST COMPONENTS
+
+define some shared fake model pieces 
+"""
+# MASK GENERATOR
+inpt = tf.keras.layers.Input((None, None, 3))
+output = tf.keras.layers.Conv2D(1, 1, padding="same", activation="sigmoid")(inpt)
+maskgen = tf.keras.Model(inpt, output)
+# INPAINTER
+inp_inpt = tf.keras.layers.Input((None, None, 4))
+output = tf.keras.layers.Conv2D(3, 1, activation="sigmoid")(inp_inpt)
+inpainter = tf.keras.Model(inp_inpt, output)
+
+
 
 
 def test_load_to_array():
@@ -16,3 +37,13 @@ def test_load_to_array():
         assert output.dtype == np.float32
         assert (output >= 0).all()
         assert (output <= 1).all()
+        
+        
+        
+def test_remove_objects(test_png_path):
+    img = Image.open(test_png_path)
+    reconstructed = _remove_objects(test_png_path, maskgen, inpainter)
+    
+    assert isinstance(reconstructed, Image.Image)
+    assert img.size == reconstructed.size
+
